+++ src/bobtemplates/gillux/nspackage/tests/__init__.py
@@ -1,6 +1,0 @@
-# -*- coding: utf-8 -*-
-"""\
-{{{ rst_pkgname }}}
-
-Tests package
-"""